delete generate and init examples
