--- benches/filter.rs
@@ -1,2 +1,0 @@
-// bloom filter vs cuckoo filter
-// in context of this project
