+++ src/util.rs
@@ -22,8 +22,10 @@ pub mod test {
     use crate::node::{NodeInfo, NodeStatus};
     use crate::node_id::NodeId;
     use crate::store::{NodeBucket, NodeTable};
+    use crate::ed25519::Keypair;
     use std::collections::VecDeque;
     use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+    use rand;
 
     pub static ADDR: &'static str = "127.0.0.1:8008";
 
@@ -37,8 +39,9 @@ pub mod test {
 
     #[test]
     fn new_node_info_succeeds() {
-        let id = NodeId::generate();
-        let new_node = new_node_info(id);
+        let key = Keypair::generate(&mut rand::thread_rng());
+        let node_id = NodeId::from_public_key(key.public);
+        let new_node = new_node_info(node_id);
         assert_eq!(new_node.port(), 8080);
         assert_eq!(new_node.ip(), IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)));
     }
@@ -50,8 +53,9 @@ pub mod test {
         };
         // should prevent duplicate NodeId generation eventually
         for i in 0..node_count {
-            let id = NodeId::generate();
-            let new_node = new_node_info(id);
+            let key = Keypair::generate(&mut rand::thread_rng());
+            let node_id = NodeId::from_public_key(key.public);
+            let new_node = new_node_info(node_id);
             bucket.nodes.push_back(new_node);
         }
         bucket
